n/a

+++ app/src/main/java/roskar/kristjan/norma/room/Norma.kt
@@ -8,8 +8,8 @@ import androidx.room.PrimaryKey
 data class Norma(
     @PrimaryKey(autoGenerate = true) val norma_id: Int?,
     @ColumnInfo(name = "datum") var norma_date: String?,
-    @ColumnInfo(name = "normaUre") var normaHours: Int?,
-    @ColumnInfo(name = "delovneUre") var workingHours: Int?,
+    @ColumnInfo(name = "normaUre") var normaHours: Double?,
+    @ColumnInfo(name = "delovneUre") var workingHours: Double?,
     @ColumnInfo(name = "delovnoMesto") var workplace: String?
 
 )
